гзв

+++ app/src/main/kotlin/com/simplemobiletools/gallery/helpers/GlideRotateTransformation.kt
@@ -7,13 +7,13 @@ import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPool
 import com.bumptech.glide.load.resource.bitmap.BitmapTransformation
 import java.security.MessageDigest
 
-class GlideRotateTransformation(context: Context, val rotateRotationAngle: Float) : BitmapTransformation(context) {
+class GlideRotateTransformation(context: Context, val rotateRotationAngle: Int) : BitmapTransformation(context) {
     override fun transform(pool: BitmapPool, bitmap: Bitmap, outWidth: Int, outHeight: Int): Bitmap {
-        if (rotateRotationAngle % 360 == 0f)
+        if (rotateRotationAngle % 360 == 0)
             return bitmap
 
         val matrix = Matrix()
-        matrix.postRotate(rotateRotationAngle)
+        matrix.postRotate(rotateRotationAngle.toFloat())
         return Bitmap.createBitmap(bitmap, 0, 0, bitmap.width, bitmap.height, matrix, true)
     }
